Extract `replace_unused_array_instructions_with_bounds_checks` optimization
