+++ compiler/noirc_evaluator/src/ssa/opt/mod.rs
@@ -17,6 +17,7 @@ mod rc;
 mod remove_bit_shifts;
 mod remove_enable_side_effects;
 mod remove_if_else;
+mod replace_unused_array_instructions_with_bounds_checks;
 mod resolve_is_unconstrained;
 mod runtime_separation;
 mod simplify_cfg;
